Reorganize project files.
