+++ core/src/main/java/org/stellar/anchor/sep12/Sep12Service.java
@@ -9,7 +9,7 @@ import org.stellar.anchor.api.sep.sep12.Sep12GetCustomerRequest;
 import org.stellar.anchor.api.sep.sep12.Sep12GetCustomerResponse;
 import org.stellar.anchor.api.sep.sep12.Sep12PutCustomerRequest;
 import org.stellar.anchor.api.sep.sep12.Sep12PutCustomerResponse;
-import org.stellar.anchor.sep10.JwtToken;
+import org.stellar.anchor.auth.JwtToken;
 import org.stellar.anchor.util.Log;
 import org.stellar.anchor.util.MemoHelper;
 import org.stellar.sdk.xdr.MemoType;
