+++ core/src/test/kotlin/org/stellar/anchor/sep10/Sep10ServiceTest.kt
@@ -32,6 +32,7 @@ import org.stellar.anchor.api.exception.SepValidationException
 import org.stellar.anchor.api.sep.sep10.ChallengeRequest
 import org.stellar.anchor.api.sep.sep10.ChallengeRequestTest
 import org.stellar.anchor.api.sep.sep10.ValidationRequest
+import org.stellar.anchor.auth.JwtService
 import org.stellar.anchor.config.AppConfig
 import org.stellar.anchor.config.Sep10Config
 import org.stellar.anchor.horizon.Horizon
