+++ platform/src/main/java/org/stellar/anchor/platform/SepConfig.java
@@ -9,8 +9,10 @@ import org.stellar.anchor.api.callback.CustomerIntegration;
 import org.stellar.anchor.api.callback.FeeIntegration;
 import org.stellar.anchor.api.callback.RateIntegration;
 import org.stellar.anchor.asset.AssetService;
+import org.stellar.anchor.auth.JwtService;
 import org.stellar.anchor.config.*;
 import org.stellar.anchor.event.EventPublishService;
+import org.stellar.anchor.filter.PlatformToAnchorTokenFilter;
 import org.stellar.anchor.filter.Sep10TokenFilter;
 import org.stellar.anchor.horizon.Horizon;
 import org.stellar.anchor.paymentservice.circle.CirclePaymentService;
@@ -20,9 +22,7 @@ import org.stellar.anchor.platform.service.ResourceReaderAssetService;
 import org.stellar.anchor.platform.service.Sep31DepositInfoGeneratorCircle;
 import org.stellar.anchor.platform.service.Sep31DepositInfoGeneratorSelf;
 import org.stellar.anchor.platform.service.SpringResourceReader;
-import org.stellar.anchor.reference.filter.PlatformToAnchorTokenFilter;
 import org.stellar.anchor.sep1.Sep1Service;
-import org.stellar.anchor.sep10.JwtService;
 import org.stellar.anchor.sep10.Sep10Service;
 import org.stellar.anchor.sep12.Sep12Service;
 import org.stellar.anchor.sep24.Sep24Service;
@@ -73,8 +73,8 @@ public class SepConfig {
     FilterRegistrationBean<PlatformToAnchorTokenFilter> registrationBean =
         new FilterRegistrationBean<>();
     registrationBean.setFilter(new PlatformToAnchorTokenFilter(anchorToPlatformJwtService));
-    registrationBean.addUrlPatterns("/transactions/*");
-    registrationBean.addUrlPatterns("/transactions");
+    //    registrationBean.addUrlPatterns("/transactions/*");
+    //    registrationBean.addUrlPatterns("/transactions");
     registrationBean.addUrlPatterns("/exchange/quotes/*");
     registrationBean.addUrlPatterns("/exchange/quotes");
     return registrationBean;
