+++ platform/src/main/java/org/stellar/anchor/platform/callback/RestCustomerIntegration.java
@@ -20,6 +20,7 @@ import org.stellar.anchor.api.sep.sep12.Sep12GetCustomerRequest;
 import org.stellar.anchor.api.sep.sep12.Sep12GetCustomerResponse;
 import org.stellar.anchor.api.sep.sep12.Sep12PutCustomerRequest;
 import org.stellar.anchor.api.sep.sep12.Sep12PutCustomerResponse;
+import org.stellar.anchor.auth.AuthHelper;
 
 public class RestCustomerIntegration implements CustomerIntegration {
   private final String anchorEndpoint;
