+++ platform/src/main/java/org/stellar/anchor/platform/controller/Sep10Helper.java
@@ -4,7 +4,7 @@ import static org.stellar.anchor.filter.BaseTokenFilter.JWT_TOKEN;
 
 import javax.servlet.http.HttpServletRequest;
 import org.stellar.anchor.api.exception.SepValidationException;
-import org.stellar.anchor.sep10.JwtToken;
+import org.stellar.anchor.auth.JwtToken;
 
 public class Sep10Helper {
   public static JwtToken getSep10Token(HttpServletRequest request) throws SepValidationException {
