+++ platform/src/main/java/org/stellar/anchor/platform/controller/Sep24Controller.java
@@ -18,8 +18,8 @@ import org.stellar.anchor.api.exception.SepNotFoundException;
 import org.stellar.anchor.api.exception.SepValidationException;
 import org.stellar.anchor.api.sep.SepExceptionResponse;
 import org.stellar.anchor.api.sep.sep24.*;
+import org.stellar.anchor.auth.JwtToken;
 import org.stellar.anchor.platform.condition.ConditionalOnAllSepsEnabled;
-import org.stellar.anchor.sep10.JwtToken;
 import org.stellar.anchor.sep24.Sep24Service;
 
 @RestController
